// src/app/core/models/invoice.ts
export interface Invoice {
    freelance: any;
    id: string,
    period: string;
    frl: string;
    file: string;
    filename: string;
    payed: string;
    payamount: string;
    amount: string;
    reference: string;
    frldisplay: string;
    sordisplay: string;
    freelancecode: string;
    stamped: string;
    estamped: string;
    mailtofrl: string;
    cancelled: string;
  }